Moves the webhook authorizer out of experimental. (#10)
* Moves the webhook authorizer out of experimental.

Since there is enough material for it to go around, it seems like a good
time to do so.  Changes, roughly:

- Renamed the package in the process, and all the mentions
of the executable name and such.

- Reapplies the Makefile formatting.  The `eval $(minikube docker-env)`
has an unfortunate issue where the trailing comment also comments out
the continuation after the `&&`.  I fixed this already in commit
3b5f92aecda863841c9eda64d837bea3c76fddf7 but it didn't get merged as of
this writing, so I just reapplied it here.

- Handled Code review comments.