Adds a first resolving authorizer implementation. (#28)
* Adds a first resolving authorizer implementation.

Note first, the approach here is wrong: we won't be reading
PolicyNodeSpecs from the API server.  That said, it seems interesting to
check out what happens when we do, so I left this in.

Changes here are:
- Factor the authorizer out of the main package.
Only server remains.

- Started resolving policy rules.

- Fix the authorizer main_test.go to pass.  Note that the current setup
*may* mean that the webhook authorizer will start denying things up
until we actually add some configuration.  But since our dev deployments
run alongside an AlwaysAllow, it may be OK.

- Added a working, but weird authorizer test.

* Handle code review comments.