Fix a resourceVersion issue in PolicyNodeWatcher.
For delete events the resourceVersion rolls backwards which can cause
us to start re-listening at the wrong version if we get disconnected.
Consequently, we can't keep track of which deletes we have received
until we get a create or modify, so we have to keep track of all deletes
seen since the last create or modify so we can prevent replay of
previous delete events.