Fixes the initialization of the API server hostport for the authorizer. (#31)
* Fixes the initialization of the API server hostport for the authorizer.

Using 'localhost' as the hostname for the API server hostport does not
work, because the apiserver certificate is not valid for localhost.  I
didn't notice this problem before because I only printed that there was
a HTTP OK status as result.  But not that the payload actually contains
an error.

Luckily, since minikube rewrites the API server certificate at start, we
can steal the information about cluster IP address and add that to
the authorizer's systemd configuration.

But, this requires that the init of systemd config be pushed until after
the minikube cluster is running, so this meant moving the systemd config
code accordingly.

In retrospect, d'oh!