Implements a hierarchical authorizer for Stolos.
The design proposal is at http://go/stolos-hierarchical-authz.

Hierarchical authorizer is a RBAC authorizer, which evaluates access
based on the roles and role bindings policies defined in the Stolos
policy hierarchy.

The general approach is to collect all roles and all role bindings
in the policy hierarchy on the path from the namespace of interest to
the policy root, and evaluate access on that.

Some questions are left open at this time, such as what happens if there
are conflicting role definitions in the policy hierarchy.  Right now this
is resolved as "higher-up role wins", but avenue is open to change this
if we decide to.

Change-Id: I5b9387dfc09f8e232340702ed77768f2bbefe62d