Fixes the behavior of Attributes.GetNamespace() in the authorizer.
The old implementation assumed that GetNamespace would never get called for a
non-resource Attribute.  However, it's not true in the RBAC authorizer.  It's not a
big deal there because its implementation of Attributes is by a struct only. So
the behavior is not documented.  However, it matters for us as we can't satisfy the
GetNamespace call if there is nowhere to get this data from.  So we introduce a special
case.

Change-Id: I701fd4dea0cbdc067bc68624b2a34db0228a077a