Implements kubectl plugin stolos get rolebindings.
Getting hierarchical roles and hierarchical role bindings are quite
similar bits of code.  This led to a unified test framework.  The two
new commands seemed to me similar enough to make almost identical and to
place in the same package.

- Adds a "fake with storage" function for the metaclient.

- Adds a new method in spacewalker.go to read the namespaces based off
of the CommandContext alone, and a test.  Refactors the test cases to
reuse common code.

Bug: 69069752
Change-Id: Idc6c1989e738a726f94ea11c668262657d72922e