Implements a flattening syncer.
Flattening syncer is a syncer binary flavor that does not rely on a
hierarchical authorizer.  Instead, the hierarchical policy information
is flattened into each supported namespace, so that the namespace
appears like a regular Kubernetes namespace.

This comes at a performance cost, but makes Stolos easier to deploy
because we should no longer need a custom hierarchical authorizer to be
added to a Kubernetes deployment with Stolos.

Bug: 70930633
Change-Id: Ic47ee20b7a580d9b47e241809ade885eee37b025