Hierarchical role and role binding evaluation.
Policy is a container for the roles and role bindings that are defined.
This is decoupled from the policyhierarchy.PolicyNode so that it can be
looked at in isolation.

PolicyNode is a node in a policy tree.  Most methods on it should be
package-private for the flattener.

PolicyTree keeps track of the policies defined so far.

The evaluation is the provisory description from:
go/stolos-role-rolebinding-evaluation

The evaluation is under discussion.  The current approach makes some
sense, so it is a fair starting point.  It should be easy to revisit
if need be.

Bug: 70930633
Change-Id: Id6d81e5ca225ef696e595644ff9d2be6b66ca8bd