Creates stolos state directory if it does not exist.
The old script just assumed it's there initially.  This commit creates
the directory if it doesn't exist, bombs out with a helpful message if
unable to act, and moves it to /usr/local/google/home/fmil/.stolos,
similar to how /usr/local/google/home/fmil/.kube is used.

Also, downgraded Kubernetes to v1.9.1 as the fastest way to unwedge
apiserver restarts.

Change-Id: I92d523b6a51ff28334412a90dab838e2b6626a76
Bug: 73441251