Factors process execution to a separate package.
The "testcontext.go" contains a bunch of reusable functionality that
can be useful outside of end-to-end tests:

- kubectl commands
- bash commands
- process execution commands

This is the first CL in a row that pulls that functionality apart and into
their own separate packages so they can be reused.  I intend to reuse
the functionality in the installer binary.

Change-Id: I778a631bdc67f1db0ad73769da73fb8221830f24