Adds test hooks into restconfig.
user.Current can not be tested in our current test harness because
it requires CGO, which our current test harness does not use.

This adds a hook that allows tests to substitute dummy current user values.
Alternative is to add CGO support into the test image, but perhaps we don't
want to do that just yet.

Change-Id: Iac841d495e07401a3a9f5a7805dc50d79586cbdd
Bug: 73886107