Makes a custom build environment for Stolos.
The change should be transparent for anyone running `make`.

The installer and libraries will need to be built with
CGO_ENABLED=1 because they depend on, say, getting the current user,
which in turn requires dynamic library builds.

I wanted to make the change directly, but it turns out it's a somewhat
more involved change.  So I'll make it separately.

Change-Id: I831dd4d397b053d8634c4be2a2045b9165c44f35