Packages the stolos installer in a container.
This allows us to deploy the installer as a stand-alone
component through docker and have a minimal download-and-install
script.

Some massaging of the startup scripts was required to get all
local directories mounted to appropriate places. At the moment
I'm more interested in validating the concept of install-via-docker
than having the installation be completely wrinkle-free.

Changes:

- The configuration now can explicitly refer to $HOME.  This was required
in order to make the files from user's home available inside the install
container.

- Added Dockerfile for building and a make target `push-installer`.
- Added code to upload the installer.
- Removed the release script -- it is no longer needed and is part of
push-installer now.
- Added an entrypoint script for the container that takes care of translating
cloud-sdk paths between host and container. (in case they differ, which
they do on my system)

Change-Id: I330a4b0047693d920161b7abd97ef1e0d04fa535