First version of the config generator.
As is, the config generator is rather useless.  It's not yet linked into
any of the installation flows.  It is partially implemented, and I think
that there is some weird problem with actually writing the configuration.

But it's a runnable skeleton and it is maybe a good idea to submit it now
first before working out the details.  This will allow followup commits
to be smaller.

It's possible to test it like so:

   $ env INTERACTIVE=1 ./run-installer.sh

where run-installer.sh is the same script as described in:

   http://go/stolos-installation

Bug: 74256459
Change-Id: I76aed1e17d7dc3a8ec567921e03cfa344536dab2