Make user's home directory path available.
This was one of the stumbling points from fishfood: in a containerized
installer the user's home directory is mounted at a fixed point (good),
but the actual path of user's $HOME is not available.

It's probably better to just make the path available, so that paths
referring to $HOME on host remain sensible in the installer too.

We do this by passing the value of $HOME into an environment variable
and then inside the container.  The installer entrypoint then symlinks
the path to /home/user to make it a valid path.

Bug: 75944028
Change-Id: Ic3959770e8b12254c5577f05cce749e2534db332