Adds cluster version check before installation.
One user feedback was due to an error which happened because their
chosen installation cluster was version 1.8.x instead 1.9.y.

This seems to indicate that even though it's pointed out in the
quickstart guide, that the mistake is easy to make for humans.  Let's
give them a nice error message in that case.

Bug: 72077917
Change-Id: I8cd2bf48fa14f4f1866c037be87ec1b5dc735761