Sets installer output dir to below .output/.../
In the old setup the installation output was dumped into the current
working directory.  This at best polluted the working directory, but at
worst could trample on useful files.

Luckily this is easy to retarget, so I did just that.  The output now
goes into a directory below our .output, which means it's out of the way
of regular work, and is cleaned up with regular 'make clean'.

Bug: 77605728
Change-Id: I74021ee26dd759169c97d06a8d54d0cd63a61c19