Passes suggested user into the batch installer.
This allows end-to-end tests to work automatically for every
tester, provided that they have cluster admin permission on the test
cluster.

In the process, I realized that configgen completely subsumes the batch
installer, and could possibly be folded into the same binary.  One
reason I'm hesitant to do so right away is that configgen has more
external dependencies.

Change-Id: Idee116d3ec2fa9fb514defec33487fcb3f2483b5
Bug: 77476659