Installer no longer barfs on permission conflict.
The old code assumed that the cluster admin permission needed for the
GKE install would always be absent from the cluster.

However, in clusters where this isn't the case, the installer would
stop.  This is wrong.  It was probably masked by us silently ignoring
errors from execution, which got fixed in
https://team-review.git.corp.google.com/c/gke-kubernetes-enterprise-control/prototype/+/240606/2.

It went unnoticed because it was masked by other brokenness of the e2e
tests, which are now also fixed.

Change-Id: I9f8178949d6d5d4561d0e2a1c43b7618d457d70a
Bug: 77723651