Use a fork of github.com/ghodss/yaml.
(My) fork at github.com/filmil/yaml has the patch that will report an
error in case unknown keys make their way into the YAML file.

The original code will silently accept unknown keys, which makes for a
confusing user experience.

The pull request https://github.com/ghodss/yaml/pull/27 that upstreams
this change is just sitting there.  But when (if?) it gets merged, we
may be able to remove this new dependency.  That is, if starts align and
it gets also used elsewhere.  Stay tuned.

Change-Id: I7dfbed86143070c7c896d3a3486218c744b02c13
Bug: 78134376