Propagates the error message from kubectl.go.
Shell invocations normally return non-informative error messages like
"exit status 1" when a subprocess fails.  Whereas, most interesting
error messaging is contained in the subprocess stderr output.

This commit embeds the stderr into the reported error.  It is not
possible to push the error report down more, since the underlying
exec.go does not have an endpoint to read the stderr, and I would not
like to buffer stderr if not necessary.

Change-Id: I6a4673f13de59984443ca3b9572c003151f8b7a2