Pass along and include installer config fields.
This introduces a new field in the config.yaml GIT_SSH_SYNC. This must
be set to false when not using SSH to sync repos.
ROOT_POLICY_DIR was renamed to POLICY_DIR, existing configs must be
updated to use the renamed field.

These are both under the `git:` map in the yaml.

Bug: 78531016
Change-Id: Ibd14c9459c36090479339a75be0077b5d74f554e