Disable GIT_KNOWN_HOSTS when known_hosts not set
git-sync supports an environment flag "GIT_KNOWN_HOSTS" to configure
whether checking against the known_hosts fils is supported. It defaults
to true, but can be set to false, which both ignores any configured
known_hosts secret and disables the "add to known hosts" question during
ssh.

I wanted this in order to support git-sync off of a the test-git-serer
within the custer, but this could be useful in other circumstances as
well and is supported by git-sync.

The interface in this version is to explicitly disable GIT_KNOWN_HOSTS
if knownHostsFilename is not set in the Nomos installer config.

Change-Id: I19d6eee2e2ad4b65d53e593bd66b3b5877631c68