Use git server s.o.t. for legacy tests
Stop applying acme.yaml directly and instead rely on the policy importer
syncing from the source of truth hosted as a git repo in the
test-git-server.

acme.yaml and the corresponding directory have diverged somewhat so a
little bit of fixing there was ncecessary to get current tests to pass.

Bug: 77476659
Change-Id: Ia60b4d375b7641947b18df6b9ae422364fc29cb2