Don't hardcode GCP_PROJECT in e2e installer
stolos-dev was hardcoded as the GCP projet for part but not
all of the test-e2e make chain. This led to problems during
release when part of the chain used GCP_PROJECT, which is
nomos-release during the chain, while the rest hardcoded stolos-dev.

Address this by using GCP_PROJECT for all of these places instead
of hardcoding.

No keys are baked into these images.

Change-Id: Id2cc451d6ecc5482ec317d468cae87ab61372a8b