Update install-config for SSH/GIT merge
We merged the ssh & git sections in the installer config,
but did not update the e2e test config.

Also, a previous change clobbered a statement that would not
put the known hosts file in the secret deploy command if it was
empty (which is is allowed to be), so put that code back.

Change-Id: I9d515c1b351a5c61e4400c97e4f1a6142d1dd3e5