Add minReadySeconds to all deployments
minReadySeconds is a deployment spec field that instructs the container liveness
prober to require that the container report ready without crashes for at least
the specified number of seconds before the pod is considered available.

In our kubectl controller, this means that we wait for the specified minimum
before we start polling for availability.

This is a k8s-builtin way to address a problem we saw specifically with the
policy-importer, where the deployment would report ready long enough that we
would report that installation had completed successfully even though the
sub-pods of the importer crashed quickly.

This change adds this field to each of the deployments (1 second is the
minimum that actually does something. We may want 0 for some deployments,
but I'm not sure which that may apply to?), and tweaks kubectl to wait for that
period before starting to poll the deployment for availability.

Change-Id: Ia24d9e4c75b02cdf00662ac38c503cd8d8ae6a86