configgen: change display of cluster contexts
The old display showed both the context name and the cluster name.  By
and large, these are the same string, and even if they aren't the
cluster name is kind of unnecessary.

Per briankennedy@'s suggestion from attached bug, the display should now
be clearer as duplication is removed and there is more room to display
the cluster name.

Screenshot here:
https://photos.app.goo.gl/J6DyJp36y2GRXRix5

Bug: 79437475
Change-Id: I218d51f2a705f5d7123f52d48d6be4c770273f53