Simplifies $HOME variable substitution.
The original code tried to do the $HOME substitution on the fly.
This noble goal clashed with the interactive installer, where it would
also replace $HOME with /user/home aftetr the user typed it in.

This is confusing since the text changes from underneath the user.

In the new approach, we only do this substitution when we actually need
to use the configuration, and the substitution is explicit.  It would
probably be an interesting exercise to formulate this in terms of an API
that is difficult to misuse, but I think the benefit for doing that as
a one-off here is negligible.  So I didn't.

Tested:
  make clean
  env NOMOS_INSTALLER_CONFIG=$HOME/nomos/config.yaml make deploy-interactive
  (verify that interactive text is untouched when edited several times)
  env NOMOS_INSTALLER_CONFIG=$HOME/nomos/config.yaml make deploy
  env NOMOS_INSTALLER_CONFIG=$HOME/nomos/config.yaml make uninstall
  make test-e2e

Bug: 79438144
Change-Id: I807a38abdc901d62d050696062714c3b05781724