Mounts input configuration as a file.
Previous code mounted the configuration's directory instead.  This required
figuring out what its directory is.  Which, in turn used 'realpath'
which is a not-exactly portable way of doing so.

Turns out (thanks to frankf@ who waved docs at me) that it is possible
to mount individual files in docker, which should be much better as we
can simply mount the file without any fiddling with the path.

This should remove the EAP issue where machines that didn't have
coreutils installed would not be able to install.  It should also make
it possible to run the installer on MacOS.

Tested: make deploy ; make test-e2e
Bug: 79438210
Change-Id: I1defbe6b27597778c9b2a1d6f2d963084a7333c8