Adds a way to translate home mounts.
Allows us to tell the installer about the user $HOME directory from the
host so that we can read and translate the references that mention
user's home dir.

This was part of EAP feedback, so I thought it would be worth
resolving.  A better approach would have been to create a symlink inside
the container, but I haven't found a good way to make it happen.

Tested: make deploy ; make test-e2e
Bug: 79438144
Change-Id: I28088278629dc08fcb429e6e236f39db39908634