Allows passing a custom CA certificate file.
This allows passing a custom CA certificate file to be used for
establishing TLS sessions in the GCP policy importer via a new flag
`--ca-file`.  The file, of course, has to exist.

In practice this is useful for passing bogus CA certificate files
generated exclusively for testing.  Using the flag it is possible to
stand up a local GFE and a local version of your API.

In theory this would also be useful for supplying alternative sources of
process identity, such as in the case of say using SPIFFE as the
certificate authority in Kubernetes.  (That said, as of this writing
that's a theoretical possibility only.)

Change-Id: I553e1f983c8571bde996d2ca250052805992c33e