Starts versioning previously unversioned files
In fishfood we found it difficult to test the installer without blessing
the release first, when there is a version skew between the
run-installer.sh script and the release.

This change makes it so that an unblessed and versioned run-installer.sh
is pushed to the release directory, which is only made visible when
release is blessed.  Same exercise is then repeated for nomosvet.sh
script and the nomos documentation.  run-installer.sh version default is
now set to be exactly the same version as the release (used to be
"stable").

Makefile.release file is now changed to allow building a dirty release.
Otherwise there isn't really a good way to *test* the release process.
But, we're still not allowed to bless a dirty release, which should be
good.

We removed the "current" artifacts, and now always release into the
"stable" sub-bucket.  The installation documentation is fixed to target
the new URL now.

Bug: 110273897
Change-Id: I371f4a6dc4a21aa721054bc42a3f9a6126f4f6c5