run-installer: Mount user's home.
This allows the user that is running the install to reach all their
files (in read only mode) that live in their home directory, using
exactly the same path as on host.  If your home dir is in
/some/dir/youruser, you are able to refer
to /some/dir/youruser/somefile.txt

While this does not solve the case of referring to arbitrary files on
host, it does solve all the specific issues that EAP users had.  The issue
in the bug quoted below was the inability to refer to a minikube
certificate file, which lives in ~/.minikube.

Other reported cases included files that lived in ~/.ssh.



Bug: 110375115

Change-Id: I8d0ed6d16994e1298dfca2de29d7b339aa635c4c