GCPPolicyImporter: allow "projects" in watcher element names.
Element names are derived from GCP resources - either folders or projects.
Importer does a sanity check that the element names look like GCP resources.
It was checking that they started either with "folders" or "namespaces".
But Watcher emits "projects", not "namespaces". This check changes importer's
expectation.

A note on why I didn't change Watcher to emit "namespaces": the change in
Watcher is a little uglier, since currently it's just copying the element
name from the resource name without looking at it. So this was easier. And I'm
not that concerned about Watcher "leaking" the fact that these resources are
projects since we're doing the same for folders already.

Bug: 110804128
Change-Id: I7e15b11a5c5a6caa55d133b08bed8915962a6ed2