Use the $HOME directory, not /tmp.
The installer script in the latest incarnation of e2e tests runs in a
docker container, which is in itself in a docker container.

So, the /tmp dir in the container is not the same as the /tmp dir in the
installer, and can't be readily mounted because run-installer.sh must
remain exactly the same as it is in regular use.

So, copy the necessary creds to $HOME, and wire through, because that
directory is accessible to both.

Bug: 111410381
Change-Id: I77cd661b83be3b75618145c30aeb46faa5487954