Adds the first GCP end-to-end test.
Adds a test fixture for GCP, and provides a single test, based on the
script at go/nomos-gcp-e2e-tests.

Automates test
https://docs.google.com/document/d/1BQ0aL8AyG3Wr7edrtMEkISbSuULOkzWoaIdAbeuHkNc/edit#bookmark=id.6nn216dvpud2

The GCP end-to-end tests require individual setup from
go/nomos-gcp-e2e-tests, and they require an existing cluster that is in
the current user's kubectl default context.  The current user is the
user that is running 'make test-e2e-gcp' i.e. on your terminal it's most
likely youraccount@google.com.

To test:
  make test-e2e-gcp
or, to avoid cleanup at the end:
  make test-e2e-noclean-gcp

or see end-to-end testing instructions for developers at:
  go/nomos-testing

Change-Id: Ide0c05d74e189974c6d82dac44442c7d177bf586