Syncer wait for importer to sync on bring-up
In order to prevent syncing an empty state, the syncer will
not start until the importer has reported that it has synced
a full state (repo snapshot or gcp batch).

This communication is side-channeled through metrics; the importer
increments an operations "sync" counter when it completes a "full
sync", and the syncer polls the metrics endpoint on the importer
service, waiting for the value to go to at least 1 before starting.

On syncer crash, as long as the importer has performed more than one
sync, the syncer will start normally. On importer crash, the syncer
will not notice through this channel unless it too crashes.

Bug: 112321191

Change-Id: I1f075bcb6fabf76360bacb4586bfcbb47916e49a