Fixes git e2e tests
Removes the 'rm -rf' of the test directory which removes files relevant
for git e2e tests.  Not sure how this came about.  git log says that the
'rm -rf' line was introduced in May, but I bet we'd have noticed
something was wrong before that.

Anyways, this seems to fix things.

Bug: 112590158
Change-Id: I46767a19db640b519cbe2d236aa755b534c9de9e