e2e: hermetic end-to-end tests
Introduces two new targets:

- make ci-test-e2e: this is a hermetic e2e test target that is suitable
for running inside a go/prow cluster.  Most code here is self-contained,
except:
  (1) the target expects that some external wrapper will check the code
out
  (2) the target expects that the credentials to use to access the test
clusters is mounted under /etc/ directory.  This is what is typically
done in a prow cluster, and is required to bootstrap the ability to use
gcloud and kubectl to begin with.

- make ci-test-e2e-local: this is a target you can run locally to run
the hermetic tests.  It downloads the credentials that the above target
would expect to be mounted.

Bug: 113165924
Change-Id: I61e427d7876c0958aeb76ba788d15bdfdd6212ba