installer: remove support for --interactive
Interactive installer mode was also apparently silently removed from the
documentation in commit 98e60cab6bd56bb12961e5623968035fc2c3323e, so
nobody knew about it anyways.

Now that we have multi-platform builds, and with imminent operator
implementation, interactive install is falling by the wayside.

Let's remove it to make our lives simpler, but first let us remove the
support for it.  This change makes it impossible to use the
--interactive flag.

Since interactivity was the last reason to keep installer dockerized,
this means we can soon drop docker as dependency.

Bug:
Change-Id: I23148a0b6c6036380632e4d4f8f5a66870343ca9