installer: dedockerized [BACKWARDS INCOMPATIBLE]
Removes docker as installer dependency, and starts using a stand-alone
installer with a minimal wrapper script.  With operator in sight,
interactive installer not used, and multi-platform builds, there is no
reason left to keep the installer in a container.

This could be further simplified similarly to how istio downloads the
minimal script, but we can do that later, and only if needed.

As a side benefit, this should make Nomos CI way less onerous.  Docker
nesting that was needed for the e2e tests was starting to get dizzying.

Changes:
- removed configgen and dependencies
- removed interactive installer code from installer
- removed docker calls from installer scripts
- removes run-installer.sh and docker call, no longer needed.
- reverts kubectl call in the entrypoint.sh script
- renames entrypoint.sh to install.sh
- fixed up e2e tests
- adjusted documentation
- introduced dependencies on gcloud and kubectl (these should be OK
  since it is unlikely that an admin would be without)

Tested:
  make -f Makefile.release release
  also, tested the install proces from the installation document.

Bug: 113185274
Change-Id: Ic51a4fd7eab6e680660c319e5cfed3161ee8c5de