Makefile.release: add "origin" remote
go/prow apparently does not define any remotes when it checks out the
git repo.  Let's define one for autorelease ourselves.  Sadly, there
seems to be nothing in the prow environment that would enable us to
avoid mentioning the repo URL explicitly, which may become a problem if
we open source this code.

Also, it may not be the only thing needed to make this work; but since
there is no way to test this change other than to deploy it on prow and
see if it succeeds, let's do that first.

Bug: 114327420
Change-Id: I5a6f46aa3b6fae01d3f5905d3143d9624c8430bb