Makefile.release: fixes the release candidates
It turns out that `git tag --sort=v:refname` does *not* sort semantic
versions correctly.  This leads to incorrectly generated next RC from a
blessed release because in the refname sorting, "1.2.3-rc.4" is
"greater" than "1.2.3", which is wrong.  Another alternative, `sort -V`
also does not sort semvers correctly (sigh), and sorts the same as git.

What to do? It turns out that while there are plenty of libraries that
parse and manage versions, none of them provide a simple utility that
sorts them.  So I wrote a small program `semsort` that does the job.
Using that program it's now possible to generate correct semvers for
releases.

In this change:

- bumped buildenv to v0.1.2 and added semsort into it for use in the
build process.

- modified the way last release candidate version is computed to use
semsort.

- tidied up the output of `make -f Makefile.release info`

Bug: 114327420
Change-Id: I030fd2cbcdb89bb58feffe7b835adf872a96f147