Makefile.release: correct the applied tag
When a tag is applied to a blessed release, it must be applied
to the same commit of the release candidate that is being blessed.

Old code applied the blessed tag to HEAD, which is wrong.  It may be
wrong because HEAD may have moved past the latest release candidate by
the time we get around to blessing a release.

Tested: git tag test HEAD^   # apply tag to one commit prior to head
  git log  # verify that the tag is indeed applied to the right commit
  git tag -d test # remove the tag
Change-Id: Ia6ac99e3574d3357e5507836437f3c810050234f