Importer: populate Resources for non-generic types
This change is not behind a flag. It doesn't stop populating the old
fields. It populates the old fields and the new ones.

This way, Generic Syncer development can proceed independently, without
requiring any breaking changes to filesystem parsing. Syncer just needs
to read the new fields and stop reading the old ones. And when that
change is finished, importer can stop populating the old fields.

Change-Id: I430df0bc58676f95b6860ac1612ed628acf5caab