Fix ssh-pub already found error
Our tests were failing with the error that ssh-pub
already existed when it should not.

We believe this happened because the workaround for
b/111218567 in setup.sh:clean_up found no policynodes,
and so skipped uninstall, but did not clean up the
test-specific resources deployed on the cluster. This
led to a loop, where the test preclean step should have
removed nomos test resources (specifically nomos-system-test
namespace), but did not, so setup couldn't create the key
because it already existed, so the tests failed because
they were using the old, incorrect key to access the remote
test git server.

The fix is to always try to clean up test resources, even
if we are not uninstalling, but do so in a way that doesn't
fail if those test resources are not found.

Bug: 116613730
Change-Id: Ia1df3582f3a3b7d51c9e7f9c08bfb70bfa2aa3db