oidc: implements an oidc command line utility
We realized that it would be beneficial to have a few handy standard
commands for interacting with OIDC clusters.

This sequence of commits implements two commands:

(1) oidc get-credentials
(2) oidc login

that are used to (1) obtain a set of creds for OIDC cluster access and
(2) to use that set of creds to log into a cluster.

Also updated dependencies. We need a bunch of new things, so let's
update them now.  I'll try to split this change out into a separate CL,
but sometimes this is simply not possible.

Change-Id: Ied2a69ac192d871e78e4629d6ee7118f1e33d386