Add ClusterObjectList and ClusterObject types.
A lot of visitor seem to be adopting a pattern where VisitObject has a
conditional for cluster vs namespace scoped objects.  This seems more
like an anti-pattern that's a consequence of the framework not
distingusihing between the two.  This fixes the anti pattern by
introducing new types so the user won't have to use potentially
non-obvious conditional for namespace vs cluster scope.

Bug: 117213141
Change-Id: I0b5abce704213cb940b14456c3310cfa0ee8d980