installer: adds a way to specify cluster name
This is not wired anywhere yet (except, if you start using it, it will
be used in the installer).

Also added a poor man's versioning for the new installer file format.
This is not as elaborate as full K8S versioning format but is *way* less
code and seemed like a good tradeoff given that we're retiring the
installer Any Day Now (tm).

There's code that shunts new fields back into old.  This should work
just fine if you use one *or* the other.

Added some docs to clarify this in the process.

Bug: 115778664
Change-Id: Id3e2606938fa8d3b47c562d85ec4d612eac2417d