installer: imports fragments of cluster-operators
These are the changes needed to get the installer the ability to work
with the same CRDs as the nomos operator.

However, since cluster-operators is a private repository at the moment,
it is not possible to vendor the dependencies in.  So we resorted to
selectively checking in parts of the source code that is needed for
the Nomos installer to support the same on-cluster CRDs.

In the process, we needed to fix a bunch of linter warnings, so the
resulting types aren't clean.

Tested:
I artificially declared a var of type addons.Nomos in installer, then
ran `make test`.

Change-Id: Ia1be484aaa3b8837a38b558644196fa84c8aae0d
Bug: 115776916