sync resources according to policynodes
This is reusing most of the old syncer code and making minor changes to
deal with generic resources. This mostly includes respecting nomos
managed labels for resources and getting resources declared in Syncs and]
embedded in the GenericResources field in PolicyNodes.

This includes unit tests for policynode reconciliation. They are mostly
targeted around GenericResource changes. It may make sense to add more
around handling namespace labeling cases.

Some monitoring was lost, since we're not using ReflectiveAction
anymore. I'll add the monitoring back in a subsequent CL.

Bug: 117236183
Change-Id: Iba5d750d86892d5acd7950d624d916a7f1bc5378