policyimporter: implements per-cluster addressing
Allows us to target resources to specific cluster selectors.  This way
different sets of resources can be deployed to different environments.

One of the rebase changes introduces ast.Root.Data type, so the code was
reworked to use that.  This made a few interventions necessary on tests
that compare the contents of ast.Root across tests, but not too bad as
can be reused once done.

Bug: 115777792
Change-Id: I51b325e6047a82b743cf200bb7a88779769824c6