Add Operator Makefile & Target
Adds an initial makefile target that downloads a "dev-stable" version of
the operator manifest from the current dev project (defaults to
stolos-dev) and applies it, resulting in more operator fun than was
available before.

Also includes a sample nomos.yaml so that the user guide can be
followed. More instructions an in-repo documentation to come.

Change-Id: I65c0c00bd0736907f7e1386bae0a1523358ce673