watch native and custom resources in syncer
I mistakenly thought you could always pass in unstructured.Unstrucutured
as the type when declaring a watch in a controller. However, this fails
when a resource has already been registered in the scheme as another
type.

This change makes it so we use the type declared in the scheme when
registering a watch.

Bug: 117236183
Change-Id: I155a7057d1f27888dbca44e90867561b641ae762