bespin: Add Bespin CRDs
This change adds the Bespin CRDs which are applied during the manifests
apply phase.

Due to a bug in GKE 1.10 Alpha, the validations are being kicked out,
which wasn't the case in base 1.10 nor in 1.11. This allows the E2E
tests to start passing again while we wait for GKE to fix.

Bug: 118383540
Change-Id: I6c6790db4655f2a07e3b938a13ef1ec7e6440cca