Additional Operator Makefile Tweaks/Support
This change updates the make process slightly for the operator to make
the cluster-operators framework happy. Specifically, we are separating
per-user dev environments by using the private-registry, but this is a
prefix only, so the suffix/tag of the image has to be something that is
already in the manifest packaged into the operator. For the "dev"
channel, this is "latest", so just use that tag.

Also update the nomos-foocorp example yaml to use the dev channel and
branch, and switch to the frankfarzan upstream fork of foo-corp-example
because I have permissions to make it up to date now 8)

Change-Id: I888cf707e4d938a8a3e2ebb332213cfc7e334bfc