update k8s libaries / enable generic syncer
This change removes the old syncer and enables the generic syncer by
default. This had to be done in one change, because the old syncer and
generic syncer can not live side by side in a functional state. The old
syncer relies on an old version of kubebuilder and the new syncer relies
on a controller-runtime version that makes it so we cannot continue to
use the old version of kubebuilder.

This change also updates the version of k8s we are using in all our k8s
related libraries to target 1.11.2 instead of 1.10.1

Some libraries from the old syncer, which are still being used have not
been removed. We should move these out of pkg/syncer at some point.

Bug: 118767144, 118189026
Change-Id: I158b06ddf8ea4effb48e7332202fe5a3ec89b15b